+++ src/core/app.rs
@@ -1,0 +1,4 @@
+pub struct App
+{
+    
+}
