+++ src/draw/bitmap.rs
