+++ src/draw/mod.rs
@@ -1,0 +1,9 @@
+mod bitmap;
+mod canvas;
+mod image;
+
+pub use self::bitmap::{ Bitmap, Buf };
+pub use self::canvas::Frame;
+pub use self::image::Image;
+
+pub use rayon::prelude::ParallelIterator;
