+++ src/math/mod.rs
@@ -1,0 +1,1 @@
+pub use vek::*;
