--- src/step.rs
@@ -1,5 +1,0 @@
-/// data passed to an update loop
-pub struct Step
-{
-
-}
